Updated URL for mempool explorer
changed from 'low' to 'medium' to 'medium' to 'high' - since I thought that it made more sense (low fees can get stuck)
Changed URL to mempool.space since its a more modern and aesthetically pleasing website. I thought the plot from bitcoinfees was hard to understand and too complicated for what should be a simple task it also has a really good graph. Also the explination that was given made sense when that text got commited but when I went to see it I thought it was really unclear.

Added that lightning network is an option to avoid all of this - was considering linking to OpenDime (by CoinKite) as well since you can make transactions (just not over the internet)

Also adding a small tip that historially sending transactions during the weekend is the best time since there is less contention for block space.